bootstrapを使用

+++ app/src/Controller/UsersController.php
@@ -11,6 +11,9 @@ use Cake\Event\Event;
  */
 class UsersController extends AppController
 {
+    public $me;
+    public $is_admin = false;
+    public $user_editable_key = ['username', 'password'];
 
     /**
      * Index method
@@ -19,12 +22,18 @@ class UsersController extends AppController
      */
     public function index()
     {
+        if (!$this->is_admin) {
+            return $this->redirect(['action'=>'edit']);
+        }
         $this->paginate = [
             'contain' => ['Groups', 'Licenses']
         ];
         $users = $this->paginate($this->Users);
 
         $this->set(compact('users'));
+        $this->set('user', $this->me);
+        $this->set('user_shop_account',
+            $this->get_user_shop_account_current($this->me->user_shop_account));
         $this->set('_serialize', ['users']);
     }
 
@@ -37,11 +46,20 @@ class UsersController extends AppController
      */
     public function view($id = null)
     {
-        $user = $this->Users->get($id, [
-            'contain' => ['Groups', 'Licenses', 'UserShopAccount', 'UserShopApplyOrderTmpl', 'UserShopEvaluateTmpl', 'UserShopReceiveFeeTmpl', 'UserShopRelistCron', 'UserShopShipTmpl']
-        ]);
+        if (!empty($id)) {
+            if (!$this->is_admin) {
+                return $this->redirect(['action'=>'edit']);
+            }
+            $user = $this->Users->get($id, [
+                'contain' => ['UserShopAccount','Groups']
+            ]);
+        } else {
+            $user = $this->me;
+        }
 
         $this->set('user', $user);
+        $this->set('user_shop_account',
+            $this->get_user_shop_account_current($user->user_shop_account));
         $this->set('_serialize', ['user']);
     }
 
@@ -52,8 +70,11 @@ class UsersController extends AppController
      */
     public function add()
     {
+        if (!$this->is_admin) {
+            return $this->redirect(['action'=>'edit']);
+        }
         $user = $this->Users->newEntity();
-        if ($this->request->is('post')) {
+        if ($this->request->is(['post', 'put'])) {
             $user = $this->Users->patchEntity($user, $this->request->data);
             if ($this->Users->save($user)) {
                 $this->Flash->success(__('The user has been saved.'));
@@ -66,6 +87,9 @@ class UsersController extends AppController
         $groups = $this->Users->Groups->find('list', ['limit' => 200]);
         $licenses = $this->Users->Licenses->find('list', ['limit' => 200]);
         $this->set(compact('user', 'groups', 'licenses'));
+        $this->set('user', $this->me);
+        $this->set('user_shop_account',
+            $this->get_user_shop_account_current($this->me->user_shop_account));
         $this->set('_serialize', ['user']);
     }
 
@@ -78,22 +102,45 @@ class UsersController extends AppController
      */
     public function edit($id = null)
     {
-        $user = $this->Users->get($id, [
-            'contain' => []
-        ]);
+        if (!empty($id)) {
+            if (!$this->is_admin) {
+                return $this->redirect(['action'=>'edit']);
+            }
+            $user = $this->Users->get($id, [
+                'contain' => ['UserShopAccount','Groups']
+            ]);
+        } else {
+            $user = $this->me;
+        }
+
         if ($this->request->is(['patch', 'post', 'put'])) {
+            if (!$this->is_admin) {
+                foreach ($this->request->data as $key => $val) {
+                    if (!in_array($key, $this->user_editable_key)) {
+                        unset($this->request->data[$key]);
+                    }
+                }
+            }
             $user = $this->Users->patchEntity($user, $this->request->data);
             if ($this->Users->save($user)) {
                 $this->Flash->success(__('The user has been saved.'));
 
-                return $this->redirect(['action' => 'index']);
+                if ($this->is_admin) {
+                    return $this->redirect(['action' => 'index']);
+                } else {
+                    return $this->redirect(['action' => 'edit']);
+                }
             } else {
                 $this->Flash->error(__('The user could not be saved. Please, try again.'));
             }
         }
+
         $groups = $this->Users->Groups->find('list', ['limit' => 200]);
         $licenses = $this->Users->Licenses->find('list', ['limit' => 200]);
         $this->set(compact('user', 'groups', 'licenses'));
+        $this->set('is_admin', $this->is_admin);
+        $this->set('user_shop_account',
+            $this->get_user_shop_account_current($user->user_shop_account));
         $this->set('_serialize', ['user']);
     }
 
@@ -106,6 +153,9 @@ class UsersController extends AppController
      */
     public function delete($id = null)
     {
+        if (!$this->is_admin) {
+            return $this->redirect(['action'=>'edit']);
+        }
         $this->request->allowMethod(['post', 'delete']);
         $user = $this->Users->get($id);
         if ($this->Users->delete($user)) {
@@ -121,6 +171,13 @@ class UsersController extends AppController
     {
         parent::beforeFilter($event);
         $this->Auth->allow(['login']);
+        $request_user = $this->Auth->user();
+        if (!empty($request_user['id'])) {
+            $this->me = $this->Users->get($request_user['id'], [
+                'contain' => ['UserShopAccount','Groups']
+            ]);
+            $this->is_admin = $this->is_admin($this->me->group);
+        }
     }
 
     public function login()
@@ -132,12 +189,35 @@ class UsersController extends AppController
                 $this->Auth->setUser($user);
                 return $this->redirect($this->Auth->redirectUrl());
             }
-            $this->Flash->error(__('Invalid username or password, try again'));
+            $this->Flash->error('ログイン情報が間違っています');
         }
     }
 
     public function logout()
     {
         return $this->redirect($this->Auth->logout());
+    }
+
+    public function get_user_shop_account_current($user_shop_accounts) {
+        if (!empty($user_shop_accounts)) {
+            if (empty($this->request->account)) {
+                return $user_shop_accounts[0];
+            } else {
+                foreach ($user_shop_accounts as $account) {
+                    if ($account->id == $this->request->account) {
+                        return $account;
+                    }
+                }
+            }
+        }
+        return null;
+    }
+
+    public function is_admin($group) {
+        if (!empty($group) && $group->name == 'admin') {
+            return true;
+        } else {
+            return false;
+        }
     }
 }
